take session ID from Authorization header
If option cookie_only is set to false and session ID cannot be found in the
cookies or in the params, then also try to get it from the Authorization header.

This makes it more convenient for mobile apps to pass the session.

The format of the header is:

```
Authorization: session_key session_value
```

So, if session key is _dawanda_session and value is abc1234, then the header
should be set to:

```
Authorization: _dawanda_session abc1234
```